+++ lib/src/widgets/md_editor_widget.dart
@@ -41,6 +41,16 @@ class MarkdownEditor extends StatefulWidget {
   /// Default value is [Colors.black]
   final Color? inputBorderColor;
 
+  /// Determines [TextFormField] cursor color
+  ///
+  /// Default value is [Colors.black]
+  final Color? inputCursorColor;
+
+  /// Determines [TextFormField] text selection color
+  ///
+  /// Default value is [Colors.black26]
+  final Color? inputTextSelectionColor;
+
   /// Determines spacing between editor and Markdown options
   ///
   /// Default value is `8`
@@ -54,6 +64,8 @@ class MarkdownEditor extends StatefulWidget {
     this.textFieldHeight = 190,
     this.textFieldMaxLines = 200,
     this.inputBorderRadius,
+    this.inputCursorColor = Colors.black,
+    this.inputTextSelectionColor,
     this.inputBorderColor = Colors.black,
     this.optionsSpacingFromEditor = 8,
   });
@@ -115,8 +127,10 @@ class _MarkdownEditorState extends State<MarkdownEditor> {
                             )
                           : TextButton(
                               style: ButtonStyle(
-                                  overlayColor: MaterialStateProperty.all(
-                                      Colors.transparent)),
+                                overlayColor: MaterialStateProperty.all(
+                                  Colors.transparent,
+                                ),
+                              ),
                               onPressed: () {
                                 controller.onSelectType(
                                     type[OptionsConstantKey.optionType]);
@@ -133,20 +147,30 @@ class _MarkdownEditorState extends State<MarkdownEditor> {
                   )
                 ],
               ),
+              // Spacing between Markdown types and text editor
               SizedBox(height: widget.optionsSpacingFromEditor),
               SizedBox(
                 height: widget.textFieldHeight,
                 width: MediaQuery.sizeOf(context).width,
-                child: TextFormField(
-                  controller: controller.textController,
-                  maxLines: widget.textFieldMaxLines,
-                  decoration: InputDecoration(
-                    border: OutlineInputBorder(
-                      borderRadius:
-                          widget.inputBorderRadius ?? BorderRadius.circular(4),
-                      borderSide: BorderSide(
-                        color: widget.inputBorderColor ?? Colors.black,
-                        width: 1.5,
+                child: Theme(
+                  data: ThemeData(
+                    textSelectionTheme: TextSelectionThemeData(
+                      cursorColor: widget.inputCursorColor,
+                      selectionColor:
+                          widget.inputTextSelectionColor ?? Colors.black26,
+                    ),
+                  ),
+                  child: TextFormField(
+                    controller: controller.textController,
+                    maxLines: widget.textFieldMaxLines,
+                    decoration: InputDecoration(
+                      border: OutlineInputBorder(
+                        borderRadius: widget.inputBorderRadius ??
+                            BorderRadius.circular(4),
+                        borderSide: BorderSide(
+                          color: widget.inputBorderColor ?? Colors.black,
+                          width: 1.5,
+                        ),
                       ),
                     ),
                   ),
@@ -159,14 +183,14 @@ class _MarkdownEditorState extends State<MarkdownEditor> {
 }
 
 /// Custom Markdown type Header Widget defined to be a
-/// [DropdownButton] for more then one kind of header type selection
+/// [DropdownButton] for more then one option of header type selection
 class _MarkdownHeaderButtonWidget extends StatelessWidget {
   /// Header options list
   ///
   /// # Header (Header 1 - #)
   /// ## Header (Header 2 - ##)
   /// ### Header (Header 3 - ###)
-  final List<String> headers;
+  final List<int> headers;
 
   /// Header option icon (already defined at [MarkdownType] enum)
   final String optionIcon;
@@ -195,28 +219,25 @@ class _MarkdownHeaderButtonWidget extends StatelessWidget {
           hoverColor: Colors.transparent,
           splashColor: Colors.transparent,
         ),
-        child: DropdownButton<String>(
+        child: DropdownButton<int>(
           underline: const SizedBox.shrink(),
           value: headers.first,
-          items: headers.map(
-            (header) {
-              final index = headers.indexOf(header);
-              return DropdownMenuItem(
-                value: header,
-                child: SvgPicture.asset(
-                  optionIcon,
-                  height: 32 - (index * 5),
-                  colorFilter: ColorFilter.mode(
-                    optionColor!,
-                    BlendMode.srcIn,
+          items: headers
+              .map(
+                (header) => DropdownMenuItem(
+                  value: header,
+                  child: SvgPicture.asset(
+                    optionIcon,
+                    height: 32 - ((header - 1) * 5),
+                    colorFilter: ColorFilter.mode(
+                      optionColor!,
+                      BlendMode.srcIn,
+                    ),
                   ),
                 ),
-              );
-            },
-          ).toList(),
-          onChanged: (value) {
-            onSelectHeader(int.parse(value!.split(' ').last));
-          },
+              )
+              .toList(),
+          onChanged: (value) => onSelectHeader(value!),
         ),
       ),
     );